requirements.txt

transformers
pillow
pydantic
fastapi
torch
uvicorn